fix: conflicts with the branch for sending only one payment methods call
