+++ src/cartridges/int_adyen_SFRA/cartridge/adyen/utils/adyenConfigs.js
@@ -139,7 +139,8 @@ const adyenConfigsObj = {
     return (
       this.isApplePayExpressEnabled() ||
       this.isAmazonPayExpressEnabled() ||
-      this.isPayPalExpressEnabled()
+      this.isPayPalExpressEnabled() ||
+      this.isGooglePayExpressEnabled()
     );
   },
 
@@ -153,6 +154,14 @@ const adyenConfigsObj = {
 
   isApplePayExpressOnPdpEnabled() {
     return getCustomPreference('ApplePayExpress_Pdp_Enabled');
+  },
+
+  isGooglePayExpressEnabled() {
+    return getCustomPreference('GooglePayExpress_Enabled');
+  },
+
+  isGooglePayExpressOnPdpEnabled() {
+    return getCustomPreference('GooglePayExpress_Pdp_Enabled');
   },
 
   isAmazonPayExpressEnabled() {
